abstract child data layers to singleton; group singletons in own asset file
