+++ assets/js/app/Singletons.js
@@ -1,0 +1,498 @@
+/* global d3,LocusZoom */
+/* eslint-env browser */
+/* eslint-disable no-console */
+
+"use strict";
+
+/**
+
+  Singletons
+
+  LocusZoom has various singleton objects that are used for registering functions or classes.
+  These objects provide safe, standard methods to redefine or delete existing functions/classes
+  as well as define new custom functions/classes to be used in a plot.
+
+*/
+
+
+/****************
+  Label Functions
+
+  These functions will generate a string based on a provided state object. Useful for dynamic axis labels.
+*/
+
+LocusZoom.LabelFunctions = (function() {
+    var obj = {};
+    var functions = {};
+
+    obj.get = function(name, state) {
+        if (!name) {
+            return null;
+        } else if (functions[name]) {
+            if (typeof state == "undefined"){
+                return functions[name];
+            } else {
+                return functions[name](state);
+            }
+        } else {
+            throw("label function [" + name + "] not found");
+        }
+    };
+
+    obj.set = function(name, fn) {
+        if (fn) {
+            functions[name] = fn;
+        } else {
+            delete functions[name];
+        }
+    };
+
+    obj.add = function(name, fn) {
+        if (functions.name) {
+            throw("label function already exists with name: " + name);
+        } else {
+            obj.set(name, fn);
+        }
+    };
+
+    obj.list = function() {
+        return Object.keys(functions);
+    };
+
+    return obj;
+})();
+
+// Label function for "Chromosome # (Mb)" where # comes from state
+LocusZoom.LabelFunctions.add("chromosome", function(state){
+    if (!isNaN(+state.chr)){ 
+        return "Chromosome " + state.chr + " (Mb)";
+    } else {
+        return "Chromosome (Mb)";
+    }
+});
+
+
+/****************
+  Scale Functions
+
+  Singleton for accessing/storing functions that will convert arbitrary data points to values in a given scale
+  Useful for anything that needs to scale discretely with data (e.g. color, point size, etc.)
+
+  All scale functions must accept an object of parameters and a value to process.
+*/
+
+LocusZoom.ScaleFunctions = (function() {
+    var obj = {};
+    var functions = {};
+
+    obj.get = function(name, parameters, value) {
+        if (!name) {
+            return null;
+        } else if (functions[name]) {
+            if (typeof parameters == "undefined" && typeof value == "undefined"){
+                return functions[name];
+            } else {
+                return functions[name](parameters, value);
+            }
+        } else {
+            throw("color function [" + name + "] not found");
+        }
+    };
+
+    obj.set = function(name, fn) {
+        if (fn) {
+            functions[name] = fn;
+        } else {
+            delete functions[name];
+        }
+    };
+
+    obj.add = function(name, fn) {
+        if (functions.name) {
+            throw("color function already exists with name: " + name);
+        } else {
+            obj.set(name, fn);
+        }
+    };
+
+    obj.list = function() {
+        return Object.keys(functions);
+    };
+
+    return obj;
+})();
+
+// Numerical Bin scale function: bin a dataset numerically by an array of breakpoints
+LocusZoom.ScaleFunctions.add("numerical_bin", function(parameters, value){
+    var breaks = parameters.breaks;
+    var values = parameters.values;
+    if (value == null || isNaN(+value)){
+        return (parameters.null_value ? parameters.null_value : values[0]);
+    }
+    var threshold = breaks.reduce(function(prev, curr){
+        if (+value < prev || (+value >= prev && +value < curr)){
+            return prev;
+        } else {
+            return curr;
+        }
+            });
+    return values[breaks.indexOf(threshold)];
+});
+
+// Categorical Bin scale function: bin a dataset numerically by matching against an array of distinct values
+LocusZoom.ScaleFunctions.add("categorical_bin", function(parameters, value){
+    if (parameters.categories.indexOf(value) != -1){
+        return parameters.values[parameters.categories.indexOf(value)];
+    } else {
+        return (parameters.null_value ? parameters.null_value : parameters.values[0]); 
+    }
+});
+
+
+/************************
+  Data Layer Subclasses
+
+  The abstract Data Layer class has general methods and properties that apply universally to all Data Layers
+  Specific data layer subclasses (e.g. a scatter plot, a line plot, gene visualization, etc.) must be defined
+  and registered with this singleton to be accessible.
+
+  All new Data Layer subclasses must be defined by accepting an id string and a layout object.
+  Singleton for storing available Data Layer classes as well as updating existing and/or registering new ones
+*/
+
+LocusZoom.DataLayers = (function() {
+    var obj = {};
+    var datalayers = {};
+
+    obj.get = function(name, id, layout) {
+        if (!name) {
+            return null;
+        } else if (datalayers[name]) {
+            if (typeof id == "undefined" || typeof layout == "undefined"){
+                throw("id or layout argument missing for data layer [" + name + "]");
+            } else {
+                return new datalayers[name](id, layout);
+            }
+        } else {
+            throw("data layer [" + name + "] not found");
+        }
+    };
+
+    obj.set = function(name, datalayer) {
+        if (datalayer) {
+            if (typeof datalayer != "function"){
+                throw("unable to set data layer [" + name + "], argument provided is not a function");
+            } else {
+                datalayers[name] = datalayer;
+                datalayers[name].prototype = new LocusZoom.DataLayer();
+            }
+        } else {
+            delete functions[name];
+        }
+    };
+
+    obj.add = function(name, datalayer) {
+        if (datalayers[name]) {
+            throw("data layer already exists with name: " + name);
+        } else {
+            obj.set(name, datalayer);
+        }
+    };
+
+    obj.list = function() {
+        return Object.keys(datalayers);
+    };
+
+    return obj;
+})();
+
+
+
+/*********************
+  Scatter Data Layer
+  Implements a standard scatter plot
+*/
+
+LocusZoom.DataLayers.add("scatter", function(id, layout){
+
+    LocusZoom.DataLayer.apply(this, arguments);
+    this.layout = layout;
+
+    // Apply defaults to the layout where missing
+    this.layout.point_size == this.layout.point_size || 4;
+    this.layout.point_shape == this.layout.point_shape || "circle";
+    this.layout.color == this.layout.color || "#888888";
+    if (this.layout.y_axis && this.layout.y_axis.axis != 1 && this.layout.y_axis.axis != 2){
+        this.layout.y_axis.axis = 1;
+    }
+
+    // Implement the main render function
+    this.render = function(){
+        this.svg.group.selectAll("*").remove(); // should this happen at all, or happen at the panel level?
+        var selection = this.svg.group
+            .selectAll("path.lz-data_layer-scatter")
+            .data(this.data)
+            .enter().append("path")
+            .attr("class", "lz-data_layer-scatter")
+            .attr("transform", function(d) {
+                var x = this.parent.state.x_scale(d[this.layout.x_axis.field]);
+                var y_scale = "y"+this.layout.y_axis.axis+"_scale";
+                var y = this.parent.state[y_scale](d[this.layout.y_axis.field]);
+                return "translate(" + x + "," + y + ")";
+            }.bind(this))
+            .attr("d", d3.svg.symbol().size(this.layout.point_size).type(this.layout.point_shape))
+            .style({ cursor: "pointer" });
+        // Apply id (if included in fields)
+        if (this.layout.fields.indexOf("id") != -1){
+            selection.attr("id", function(d){ return d.id; });
+        }
+        // Apply color
+        if (this.layout.color){
+            switch (typeof this.layout.color){
+                case "string":
+                    selection.attr("fill", this.layout.color);
+                    break;
+                case "object":
+                    if (this.layout.color.scale_function && this.layout.color.field) {
+                        selection.attr("fill", function(d){
+                            return LocusZoom.ScaleFunctions.get(this.layout.color.scale_function,
+                                                                this.layout.color.parameters || {},
+                                                                d[this.layout.color.field]);
+                        }.bind(this));
+                    }
+                    break;
+            }
+        }
+        // Apply title (basic mouseover label)
+        if (this.layout.point_label_field){
+            selection.append("svg:title")
+                .text(function(d) { return d[this.layout.point_label_field]; }.bind(this));
+        }
+    };
+       
+    return this;
+});
+
+/*********************
+  Genes Data Layer
+  Implements a data layer that will render gene tracks
+*/
+
+LocusZoom.DataLayers.add("genes", function(id, layout){
+
+    LocusZoom.DataLayer.apply(this, arguments);
+    this.layout = layout;
+
+    // Apply defaults to the layout where missing
+    this.layout.track_height = this.layout.track_height || 40;
+    this.layout.label_font_size = this.layout.label_font_size || 12;
+    this.layout.track_vertical_spacing = this.layout.track_vertical_spacing || 8;
+    this.layout.label_vertical_spacing = this.layout.label_vertical_spacing || 4;
+    
+    this.metadata.tracks = 1;
+    this.metadata.gene_track_index = { 1: [] }; // track-number-indexed object with arrays of gene indexes in the dataset
+    this.metadata.horizontal_padding = 4; // pixels to pad on either side of a gene or label when determining collisions
+
+    // After we've loaded the genes interpret them to assign
+    // each to a track so that they do not overlap in the view
+    this.assignTracks = function(){
+
+        // Function to get the width in pixels of a label given the text and layout attributes
+        this.getLabelWidth = function(gene_name, font_size){
+            var temp_text = this.svg.group.append("text")
+                .attr("x", 0).attr("y", 0).attr("class", "lz-gene lz-label")
+                .style("font-size", font_size)
+                .text(gene_name + "→");
+            var label_width = temp_text.node().getBBox().width;
+            temp_text.node().remove();
+            return label_width;
+        };
+
+        // Reinitialize metadata
+        this.metadata.tracks = 1;
+        this.metadata.gene_track_index = { 1: [] };
+
+        this.data.map(function(d, g){
+
+            // Determine display range start and end, based on minimum allowable gene display width, bounded by what we can see
+            // (range: values in terms of pixels on the screen)
+            this.data[g].display_range = {
+                start: this.parent.state.x_scale(Math.max(d.start, this.parent.parent.state.start)),
+                end:   this.parent.state.x_scale(Math.min(d.end, this.parent.parent.state.end))
+            };
+            this.data[g].display_range.label_width = this.getLabelWidth(this.data[g].gene_name, this.layout.label_font_size);
+            this.data[g].display_range.width = this.data[g].display_range.end - this.data[g].display_range.start;
+            this.data[g].display_range.text_anchor = "middle";
+            if (this.data[g].display_range.width < this.data[g].display_range.label_width){
+                if (d.start < this.parent.parent.state.start){
+                    this.data[g].display_range.end = this.data[g].display_range.start
+                        + this.data[g].display_range.label_width
+                        + this.metadata.horizontal_padding;
+                    this.data[g].display_range.text_anchor = "start";
+                } else if (d.end > this.parent.parent.state.end){
+                    this.data[g].display_range.start = this.data[g].display_range.end
+                        - this.data[g].display_range.label_width
+                        - this.metadata.horizontal_padding;
+                    this.data[g].display_range.text_anchor = "end";
+                } else {
+                    var centered_margin = ((this.data[g].display_range.label_width - this.data[g].display_range.width) / 2)
+                        + this.metadata.horizontal_padding;
+                    if ((this.data[g].display_range.start - centered_margin) < this.parent.state.x_scale(this.parent.parent.state.start)){
+                        this.data[g].display_range.start = this.parent.state.x_scale(this.parent.parent.state.start);
+                        this.data[g].display_range.end = this.data[g].display_range.start + this.data[g].display_range.label_width;
+                        this.data[g].display_range.text_anchor = "start";
+                    } else if ((this.data[g].display_range.end + centered_margin) > this.parent.state.x_scale(this.parent.parent.state.end)) {
+                        this.data[g].display_range.end = this.parent.state.x_scale(this.parent.parent.state.end);
+                        this.data[g].display_range.start = this.data[g].display_range.end - this.data[g].display_range.label_width;
+                        this.data[g].display_range.text_anchor = "end";
+                    } else {
+                        this.data[g].display_range.start -= centered_margin;
+                        this.data[g].display_range.end += centered_margin;
+                    }
+                }
+                this.data[g].display_range.width = this.data[g].display_range.end - this.data[g].display_range.start;
+            }
+            // Convert and stash display range values into domain values
+            // (domain: values in terms of the data set, e.g. megabases)
+            this.data[g].display_domain = {
+                start: this.parent.state.x_scale.invert(this.data[g].display_range.start),
+                end:   this.parent.state.x_scale.invert(this.data[g].display_range.end)
+            };
+            this.data[g].display_domain.width = this.data[g].display_domain.end - this.data[g].display_domain.start;
+
+            // Using display range/domain data generated above cast each gene to tracks such that none overlap
+            this.data[g].track = null;
+            var potential_track = 1;
+            while (this.data[g].track == null){
+                var collision_on_potential_track = false;
+                this.metadata.gene_track_index[potential_track].map(function(placed_gene){
+                    if (!collision_on_potential_track){
+                        var min_start = Math.min(placed_gene.display_range.start, this.display_range.start);
+                        var max_end = Math.max(placed_gene.display_range.end, this.display_range.end);
+                        if ((max_end - min_start) < (placed_gene.display_range.width + this.display_range.width)){
+                            collision_on_potential_track = true;
+                        }
+                    }
+                }.bind(this.data[g]));
+                if (!collision_on_potential_track){
+                    this.data[g].track = potential_track;
+                    this.metadata.gene_track_index[potential_track].push(this.data[g]);
+                } else {
+                    potential_track++;
+                    if (potential_track > this.metadata.tracks){
+                        this.metadata.tracks = potential_track;
+                        this.metadata.gene_track_index[potential_track] = [];
+                    }
+                }
+            }
+
+            // Stash parent references on all genes, trascripts, and exons
+            this.data[g].parent = this;
+            this.data[g].transcripts.map(function(d, t){
+                this.data[g].transcripts[t].parent = this.data[g];
+                this.data[g].transcripts[t].exons.map(function(d, e){
+                    this.data[g].transcripts[t].exons[e].parent = this.data[g].transcripts[t];
+                }.bind(this));
+            }.bind(this));
+
+        }.bind(this));
+        return this;
+    };
+
+    // Implement the main render function
+    this.render = function(){
+
+        this.assignTracks();
+
+        this.svg.group.selectAll("*").remove();
+
+        // Render gene groups
+        this.svg.group.selectAll("g.lz-gene").data(this.data).enter()
+            .append("g")
+            .attr("class", "lz-gene")
+            .attr("id", function(d){ return d.gene_name; })
+            .each(function(gene){
+
+                // Render gene boundaries
+                d3.select(this).selectAll("rect.lz-gene").filter(".lz-boundary")
+                    .data([gene]).enter().append("rect")
+                    .attr("class", "lz-gene lz-boundary")
+                    .attr("id", function(d){ return d.gene_name; })
+                    .attr("x", function(d){ return this.parent.state.x_scale(d.start); }.bind(gene.parent))
+                    .attr("y", function(d){
+                        var exon_height = this.parent.layout.track_height
+                            - this.parent.layout.track_vertical_spacing
+                            - this.parent.layout.label_font_size
+                            - this.parent.layout.label_vertical_spacing;
+                        return ((d.track-1) * this.parent.layout.track_height)
+                            + (this.parent.layout.track_vertical_spacing / 2)
+                            + this.parent.layout.label_font_size
+                            + this.parent.layout.label_vertical_spacing
+                            + (Math.max(exon_height, 3) / 2);
+                    }.bind(gene)) // Arbitrary track height; should be dynamic
+                    .attr("width", function(d){ return this.parent.state.x_scale(d.end) - this.parent.state.x_scale(d.start); }.bind(gene.parent))
+                    .attr("height", 1) // This should be scaled dynamically somehow
+                    .attr("fill", "#000099")
+                    .style({ cursor: "pointer" })
+                    .append("svg:title")
+                    .text(function(d) { return d.gene_name; });
+
+                // Render gene labels
+                d3.select(this).selectAll("text.lz-gene")
+                    .data([gene]).enter().append("text")
+                    .attr("class", "lz-gene lz-label")
+                    .attr("x", function(d){
+                        if (d.display_range.text_anchor == "middle"){
+                            return d.display_range.start + (d.display_range.width / 2);
+                        } else if (d.display_range.text_anchor == "start"){
+                            return d.display_range.start;
+                        } else if (d.display_range.text_anchor == "end"){
+                            return d.display_range.end;
+                        }
+                    })
+                    .attr("y", function(d){
+                        return ((d.track-1) * this.parent.layout.track_height)
+                            + (this.parent.layout.track_vertical_spacing / 2)
+                            + this.parent.layout.label_font_size;
+                    }.bind(gene))
+                    .attr("text-anchor", function(d){ return d.display_range.text_anchor; })
+                    .style("font-size", gene.parent.layout.label_font_size)
+                    .text(function(d){ return (d.strand == "+") ? d.gene_name + "→" : "←" + d.gene_name; });
+
+                // Render exons (first transcript only, for now)
+                d3.select(this).selectAll("g.lz-gene").filter(".lz-exons")
+                    .data([gene]).enter().append("g")
+                    .attr("class", "lz-gene lz-exons")
+                    .each(function(gene){
+
+                        d3.select(this).selectAll("rect.lz-gene").filter(".lz-exon")
+                            .data(gene.transcripts[0].exons).enter().append("rect")
+                            .attr("class", "lz-gene lz-exon")
+                            .attr("id", function(d){ return d.exon_id; })
+                            .attr("x", function(d){ return this.parent.state.x_scale(d.start); }.bind(gene.parent))
+                            .attr("y", function(){
+                                return ((this.track-1) * this.parent.layout.track_height)
+                                    + (this.parent.layout.track_vertical_spacing / 2)
+                                    + this.parent.layout.label_font_size
+                                    + this.parent.layout.label_vertical_spacing;
+                            }.bind(gene))
+                            .attr("width", function(d){
+                                return this.parent.state.x_scale(d.end) - this.parent.state.x_scale(d.start);
+                            }.bind(gene.parent))
+                            .attr("height", function(d){
+                                return this.parent.layout.track_height
+                                    - this.parent.layout.track_vertical_spacing
+                                    - this.parent.layout.label_font_size
+                                    - this.parent.layout.label_vertical_spacing;
+                            }.bind(gene))
+                            .attr("fill", "#000099")
+                            .style({ cursor: "pointer" });
+
+                    });
+
+            });
+        
+    };
+       
+    return this;
+});
