+++ gulpfile.js
@@ -14,7 +14,8 @@ var app_js_files = ["./assets/js/app/LocusZoom.js",
                     "./assets/js/app/Data.js",
                     "./assets/js/app/Instance.js",
                     "./assets/js/app/Panel.js",
-                    "./assets/js/app/DataLayer.js"
+                    "./assets/js/app/DataLayer.js",
+                    "./assets/js/app/Singletons.js"
                    ];
 
 // Test suites. Should be able to be executed in any order.
@@ -22,7 +23,8 @@ var test_js_files = ["./test/LocusZoom.js",
                      "./test/Data.js",
                      "./test/Instance.js",
                      "./test/Panel.js",
-                     "./test/DataLayer.js"
+                     "./test/DataLayer.js",
+                     "./test/Singletons.js"
                     ];
 
 // Vendor libraries. These are *only* libraries necessary for implementing the plugin.
